指标管理，用户可以查看指定年度本级机构和下级机构 对于像长春市这样不处理业务的机构，直接查询下级机构的指标 删除校验 修改校验，及处理 增加了发放指标的机构id，便于查询

+++ src/main/java/com/cc/buildingReform/service/QuotaServiceImpl.java
@@ -1,12 +1,15 @@
 package com.cc.buildingReform.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.cc.buildingReform.dao.DepartmentDAO;
 import com.cc.buildingReform.dao.QuotaDAO;
+import com.cc.buildingReform.form.Department;
 import com.cc.buildingReform.form.Quota;
 import com.cc.buildingReform.form.User;
 
@@ -17,6 +20,9 @@ public class QuotaServiceImpl implements QuotaService {
 	@Autowired
 	private QuotaDAO quotaDAO;
 
+	@Autowired
+	private DepartmentDAO departmentDAO;
+
 	public void save(Quota quota) {
 		if (quota.getId() != null && quota.getId() != 0) {
 			// 校验更改后的指标数量和剩余数量是否符合规则
@@ -25,6 +31,24 @@ public class QuotaServiceImpl implements QuotaService {
 			if (restNum < 0) {
 				throw new RuntimeException("-1");
 			}
+			
+			// 不是省厅用户修改自己的指标
+			if (!quota.getDepartmentId().equals("01")) {
+				// 校验 新修改的下属机构指标 和自己剩余指标数量是否匹配
+				List<Quota> list = quotaDAO.findByDepartmentId(oldQuota.getYear(), oldQuota.getDistributeDepartmentId());
+				if (list == null || list.isEmpty()) {
+					throw new RuntimeException("-3");
+				}
+				
+				Integer selfRestNum = list.get(0).getRestNum() - (quota.getNum() - oldQuota.getNum());
+				if (selfRestNum < 0) {
+					throw new RuntimeException("-4");
+				}
+				
+				list.get(0).setRestNum(selfRestNum);
+				
+				quotaDAO.update(list.get(0));
+			} 
 			
 			oldQuota.setNum(quota.getNum());
 			oldQuota.setRestNum(restNum);
@@ -52,17 +76,21 @@ public class QuotaServiceImpl implements QuotaService {
 				throw new RuntimeException("-1");
 			}
 			
+			// 本单位指标 - 1
 			list.get(0).setRestNum(list.get(0).getRestNum() - quota.getNum());
 			quotaDAO.saveOrUpdate(list.get(0));
 			
 			quota.setRestNum(quota.getNum());
-			
+			quota.setDistributeDepartmentId(user.getDepartmentId());
 			quotaDAO.saveOrUpdate(quota);
 		}
 	}
 	
 	public void remove(Integer id) {
-		// TODO 删除前校验此项年度指标是否已经使用，如果使用不允许删除 
+		Quota quota = quotaDAO.get(id);
+		if (!quota.getNum().equals(quota.getRestNum())) {
+			throw new RuntimeException("-1");
+		}
 		
 		quotaDAO.delete(id);
 	}
@@ -89,5 +117,18 @@ public class QuotaServiceImpl implements QuotaService {
 	
 	public List<Integer> findExistYear() {
 		return quotaDAO.findExistYear();
+	}
+	
+	public List<Quota> findByFatherDepartmentId(Integer year, String fatherDepartmentId) {
+		Department department = departmentDAO.get(fatherDepartmentId);
+		
+		if (department.getIsWork() == 1) {
+			// 对于处理业务的机构，有上级下发的指标，同时也有给下级下发的指标
+			return quotaDAO.findByFatherDepartmentId(year, fatherDepartmentId);
+		} else {
+			// 对于不处理业务的机构，没有上级下发的指标，但下级肯定会有指标
+			return quotaDAO.findByDepartmentId(year, departmentDAO.findByFatherId(fatherDepartmentId)
+					.stream().map(Department::getId).collect(Collectors.toList()));
+		}
 	}
 }
